fix cardano buttons view
