+++ src/components/SocialLinks/SocialLinks.less
@@ -12,6 +12,10 @@
     margin-right: 4px;
   }
 
+  &__leave-feedback {
+    margin-left: 4px;
+  }
+
   &__link {
     color: var(--ergo-primary-color);
